updated, changed button drawables

+++ RHRSudoku/src/com/example/rhrsudoku/SudokuPuzzleCell.java
@@ -57,11 +57,11 @@ public class SudokuPuzzleCell {
 		return inputMethod;
 	}
 	
-	public int getValue() {
+	public Integer getValue() {
 		if (this.hasValue)
 			return value;
 		else
-			return (Integer) null;
+			return null;
 	}
 	
 	public void setNeighbours(SudokuPuzzleCell[] neighbours) {
